Create PRODIGY_WD_04
The Personal Portfolio Website is a professionally designed, responsive web application created using HTML, CSS, and JavaScript. It showcases a web developer’s skills, projects, background, and contact information in an elegant and organized manner.

This website serves as an online digital resume and a personal brand platform, designed to leave a strong impression on recruiters, clients, or collaborators by providing a visually appealing and interactive browsing experience.